pyproject.toml
[project]
name = "geovault"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crispy-bootstrap5>=2025.4",
    "django>=5.2",
    "django-crispy-forms>=2.3",
    "django-extensions>=4.1",
    "django-geojson>=4.2.0",
    "django-leaflet>=0.31.0",
    "django-role-permissions>=3.2.0",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "python-decouple>=3.8",
    "ruff>=0.11.5",
]


[tool.ruff]
line-length = 120
target-version = "py312"
output-format = "grouped"
lint.extend-select = [
    "I001",   # unsorted-imports
    "RUF019", # unnecessary-key-check
    "RUF100", # unused-noqa
    "UP006",  # non-pep585-annotation
    "UP007",  # pep604 int | None = None instead of using Optional.
]

[tool.pytest.ini_options]
minversion = "8.3.5"
addopts = "-ra -q"
testpaths = ["tests"]
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=9.1.0",
]
